// types/project.ts
export interface Project {
  title: string;
  description: string;
  technologies: string[];
  stats?: string;
  images: string[];
  video?: string;
  challenges: string[];
  solutions: string[];
  impact: string[];
  demoUrl?: string;
  githubUrl?: string;
  timeline?: string;
  role?: string;
  team?: string[];
}